refactor(cli): extract recall command family into cli/commands/recall.ts
